Moving NPartnersAnalyzer into stisim/analyzers.py
