+++ stisim/analyzers.py
@@ -888,3 +888,157 @@ class PartnershipFormationAnalyzer(ss.Analyzer):
         for i, (lo, hi) in enumerate(self.age_ranges):
             counts[i] = int(np.count_nonzero(sex_mask & (ages >= lo) & (ages < hi)))
         return counts
+
+
+class NPartnersAnalyzer(ss.Analyzer):
+    """Track partnerships per agent over a trailing month window.
+
+    Works with any male-female sexual network (e.g. ``MFNetwork``,
+    ``SWNetwork``, ``StructuredSexual``), which all share the edge convention
+    that ``p1`` is male and ``p2`` is female. Same-sex networks are not
+    supported.
+
+    Walks the active network edges each step, recording (per timestep) the
+    male/female partners and each edge's ``formation_ti``. After the sim
+    finishes:
+
+      - ``get_unique_partner_counts(female)`` returns the number of *unique*
+        opposite-sex partners each agent saw in the trailing window.
+      - ``get_n_partnerships_formed(female)`` returns the *total* number of
+        partnerships each agent formed in the trailing window (NOT unique:
+        forming three partnerships with the same person counts as three).
+
+    Both methods toggle their subject population with the ``female`` arg:
+    ``False`` (default) → results per male, ``True`` → results per female.
+
+    Args:
+        network (str): Network name to inspect (default ``'mfnetwork'``).
+        window_months (int): Trailing window in months; assumes monthly dt
+            (one ti per month) so this is also the number of trailing steps.
+    """
+
+    def __init__(self, network='mfnetwork', window_months=12, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.network = network
+        self.window_months = window_months
+        # list of (ti, male_uids, female_uids, formation_ti, age_male, age_female)
+        # where age_* are each partner's age recorded on the edge at formation.
+        self._records = []
+        self._final_male_uids = None
+        self._final_female_uids = None
+        self._final_ti = None
+        return
+
+    def step(self):
+        sim = self.sim
+        nw = sim.networks[self.network]
+        p1 = np.asarray(nw.edges.p1, dtype=np.int64)
+        p2 = np.asarray(nw.edges.p2, dtype=np.int64)
+        formation_ti = np.asarray(nw.edges.formation_ti, dtype=np.int64)
+        # age_p1 / age_p2 are stored on the edge at formation, so they give each
+        # partner's age at the moment the partnership formed (not the current ti).
+        age_p1 = np.asarray(nw.edges.age_p1, dtype=float)
+        age_p2 = np.asarray(nw.edges.age_p2, dtype=float)
+        if len(p1):
+            self._records.append((self.ti, p1, p2, formation_ti, age_p1, age_p2))
+        # Cache so we can compute counts even after sim is finalized/shrunk
+        self._final_male_uids = np.asarray(sim.people.male.uids, dtype=np.int64)
+        self._final_female_uids = np.asarray(sim.people.female.uids, dtype=np.int64)
+        self._final_ti = self.ti
+        return
+
+    def _subject_uids(self, female):
+        """Return the cached uid array for the requested subject population."""
+        return self._final_female_uids if female else self._final_male_uids
+
+    def get_unique_partner_counts(self, female=False):
+        """Return per-agent unique opposite-sex partner counts over the window.
+
+        Args:
+            female (bool): If False (default), count unique female partners per
+                male. If True, count unique male partners per female.
+
+        Returns:
+            ndarray indexed like the subject-population uid array (males if
+            ``female`` is False, females if True).
+        """
+        threshold = self._final_ti - self.window_months + 1
+        partner_sets = defaultdict(set)
+        for ti, males, females, fti, age_m, age_f in self._records:
+            if ti < threshold:
+                continue
+            # p1 is male, p2 is female. Key on the subject sex and accumulate
+            # the (unique) partner of the other sex.
+            if female:
+                for m, f in zip(males, females):
+                    partner_sets[int(f)].add(int(m))
+            else:
+                for m, f in zip(males, females):
+                    partner_sets[int(m)].add(int(f))
+
+        uids = self._subject_uids(female)
+        counts = np.zeros(len(uids), dtype=np.int64)
+        for i, uid in enumerate(uids):
+            counts[i] = len(partner_sets.get(int(uid), ()))
+        return counts
+
+    def get_n_partnerships_formed(self, female=False, age_bins=None):
+        """Return the count of partnerships formed in the trailing window.
+
+        Unlike :meth:`get_unique_partner_counts`, this is a *non-unique* count:
+        if an agent forms three partnerships with the same partner during the
+        window, all three are counted. Formation events are identified from
+        the edges' ``formation_ti`` (an edge is "formed at" a step when its
+        ``formation_ti`` equals that step's ``ti``).
+
+        Args:
+            female (bool): If False (default), count partnerships formed by
+                males. If True, count by females.
+            age_bins (list[str] | None): If None (default), return a per-agent
+                count indexed like the subject-population uid array. If a list
+                of age-range strings parsable by ``ss.parse_age_range`` (e.g.
+                ``['0-5', '5-10', ...]``), return a per-bin count where each
+                formed partnership is attributed to every bin containing the
+                subject's age *at formation* (so overlapping bins each count
+                it). The bin age uses the partner's age recorded on the edge.
+
+        Returns:
+            ndarray. If ``age_bins`` is None: indexed like the subject-population
+            uid array. Otherwise: length ``len(age_bins)``, ordered as given.
+        """
+        threshold = self._final_ti - self.window_months + 1
+
+        if age_bins is None:
+            counts_by_uid = defaultdict(int)
+            for ti, males, females, fti, age_m, age_f in self._records:
+                if ti < threshold:
+                    continue
+                # Edges newly formed at this step (counted once, at formation).
+                new_mask = fti == ti
+                if not new_mask.any():
+                    continue
+                subjects = females[new_mask] if female else males[new_mask]
+                for uid in subjects:
+                    counts_by_uid[int(uid)] += 1
+
+            uids = self._subject_uids(female)
+            counts = np.zeros(len(uids), dtype=np.int64)
+            for i, uid in enumerate(uids):
+                counts[i] = counts_by_uid.get(int(uid), 0)
+            return counts
+
+        # Binned by the subject's age at formation. Overlapping bins each count.
+        ranges = [ss.parse_age_range(b) for b in age_bins]
+        counts = np.zeros(len(ranges), dtype=np.int64)
+        for ti, males, females, fti, age_m, age_f in self._records:
+            if ti < threshold:
+                continue
+            new_mask = fti == ti
+            if not new_mask.any():
+                continue
+            ages = (age_f if female else age_m)[new_mask]
+            for a in ages:
+                for bi, (lo, hi) in enumerate(ranges):
+                    if lo <= a < hi:
+                        counts[bi] += 1
+        return counts
